fixed pmd violations

+++ src/main/java/edu/depaul/se433/shoppingapp/PurchaseAgent.java
@@ -37,9 +37,9 @@ public class PurchaseAgent {
 
   public double averagePurchase(String user) {
     List<Purchase> purchases = dbo.getPurchases(user);
-    short cnt = 0;
+    int cnt = 0;
     double total = 0.0;
-    for (short i = 0; i < purchases.size(); i++) {
+    for (int i = 0; i < purchases.size(); i++) {
       Purchase p = purchases.get(i);
       cnt++;
       total += p.getCost();
